Bug fixed: error when updating a product

+++ src/main/java/dev/bogdanbalalau/onlinestorebackend/entity/Product.java
@@ -22,5 +22,5 @@ public class Product {
 
     private Double price;
 
-    private Integer size;
+    private Double size;
 }
